+++ src/App.js
@@ -1,9 +1,9 @@
 import { Container, CssBaseline } from "@mui/material";
 import React from "react";
-import DataTable from "./components/DataTable";
+import DataTable from "./components/dataTable/DataTable";
 import { Header } from "./components/global/Header";
 import { AppProvider } from "./context/AppContext";
-import { theme } from "./global/theme";
+import { theme } from "./theme";
 import { ThemeProvider } from "@mui/material/styles";
 import { MetaTags } from "react-meta-tags";
